Touchup commit for adding UDP support
This commit has a variety of cleanups/touchups to the add-udp-support
stack of commits:
    * Renamed Raven::Transport to Raven::Transports
    * Fixed Ruby 1.9 hash usage in a couple places
    * Removed the use of .tap in favor of a 'begin/end' block
    * More widespread use of configuration.value rather than
      configuration[:value] per code review
    * Other minor cleanups
